+++ src/Traits/WithPagination.php
@@ -39,48 +39,34 @@ trait WithPagination
 
     protected bool $shouldShowPaginationDetails = true;
 
-    // TODO: Test
-    public function setupPagination(): void
-    {
-        if ($this->paginationIsDisabled()) {
-            return;
-        }
-
-        if (in_array(session($this->getPerPagePaginationSessionKey(), $this->getPerPage()), $this->getPerPageAccepted(), true)) {
-            $this->setPerPage(session($this->getPerPagePaginationSessionKey(), $this->getPerPage()));
-        } else {
-            $this->setPerPage($this->getPerPageAccepted()[0] ?? 10);
-        }
-    }
 
     // TODO: Test
     public function updatedPerPage($value): void
     {
-        if (! in_array($value, $this->getPerPageAccepted(), false)) {
-            $value = $this->setPerPage($this->getPerPageAccepted()[0] ?? 10);
+        if (! in_array((int) $value, $this->getPerPageAccepted(), false)) {
+            $value = $this->getPerPageAccepted()[0] ?? 10;
         }
 
-        if (in_array(session($this->getPerPagePaginationSessionKey(), $this->getPerPage()), $this->getPerPageAccepted(), true)) {
+        if (in_array(session($this->getPerPagePaginationSessionKey(), (int) $value), $this->getPerPageAccepted(), true)) {
             session()->put($this->getPerPagePaginationSessionKey(), (int) $value);
         } else {
             session()->put($this->getPerPagePaginationSessionKey(), $this->getPerPageAccepted()[0] ?? 10);
         }
-
-        $this->resetComputedPage();
+        $this->setPerPage($value);
     }
 
-    /**
-     * Reset the page using the custom page name
-     */
-    public function resetComputedPage(): void
+    protected function queryStringWithPagination(): array
     {
-        $this->resetPage($this->getComputedPageName());
+
+        if ($this->queryStringIsEnabled()) {
+            return [
+                'perPage' => ['except' => null, 'history' => false, 'keep' => false, 'as' => $this->getQueryStringAlias().'perPage'],
+            ];
+        }
+
+        return [];
     }
 
-    private function getPerPagePaginationSessionKey(): string
-    {
-        return $this->tableName.'-perPage';
-    }
 
     public function renderingWithPagination(): void
     {