+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to `laravel-livewire-tables` will be documented in this file
 ## UNRELEASED
 ### New Features
 - Add capability to show/hide Pagination Details (Rows x of y) by @lrljoe in https://github.com/rappasoft/laravel-livewire-tables/pull/1618
+- Add perPage to queryString by @lrljoe in https://github.com/rappasoft/laravel-livewire-tables/pull/1618
 
 ## [v3.1.7] - 2023-12-29
 ### Bug Fixes